pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "whatsfinance"
version = "0.2.0"
description = "Bot Telegram + dashboard web para finanças pessoais (Supabase)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask",
    "requests",
    "supabase",
    "python-dotenv",
    "gunicorn",
    "google-generativeai",
    "Pillow",
    "python-dateutil",
    "flask-login",
    "werkzeug",
    "email_validator",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["whatsfinance*"]
